:recycle: refactor: restructure exception package by domain
                - Moved authentication-related exceptions to exception.auth
                - Moved data-related exceptions (e.g., duplication, not found) to exception.data
                - Organized exception structure for better readability and maintenance
