+++ src/main/java/app/handong/feed/exception/data/NotFoundException.java
@@ -1,4 +1,4 @@
-package app.handong.feed.exception;
+package app.handong.feed.exception.data;
 
 import lombok.NoArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -10,10 +10,10 @@ import org.springframework.web.bind.annotation.ResponseStatus;
  * 해당 메서드 권한 없을 경우 사용되는 예외처리
  * HttpStatus FORBIDDEN
  */
-@ResponseStatus(value = HttpStatus.UNAUTHORIZED)
+@ResponseStatus(value = HttpStatus.NOT_FOUND)
 @NoArgsConstructor
-public class NoAuthenticatedException extends RuntimeException {
-    public NoAuthenticatedException(String message) {
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
         super(message);
     }
 }
